refactor controller #29
