+++ app/chatbot/messagetemplates/favorites.py
